+++ drizzle/20260612112659_fat_wallop/migration.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `prompts` (
-	`id` integer PRIMARY KEY AUTOINCREMENT NOT NULL,
+	`id` integer PRIMARY KEY AUTOINCREMENT,
 	`clientName` text NOT NULL,
-	`requestId` integer NOT NULL,
+	`requestId` text NOT NULL,
 	`callbackUrl` text,
 	`callbackCompleted` integer NOT NULL,
 	`createdAt` integer NOT NULL,
@@ -11,6 +11,7 @@ CREATE TABLE `prompts` (
 	`systemPrompt` text,
 	`userPrompt` text NOT NULL,
 	`temperature` real NOT NULL,
+	`priority` integer DEFAULT 0 NOT NULL,
 	`retryCount` integer NOT NULL,
 	`nextRetryAt` integer,
 	`reasoning` text,
@@ -21,7 +22,7 @@ CREATE TABLE `prompts` (
 	`responseTokenPerSecond` integer
 );
 --> statement-breakpoint
-CREATE INDEX `idx_prompts_callback` ON `prompts` (`callbackCompleted`,`status`,`callbackUrl`);--> statement-breakpoint
-CREATE INDEX `idx_prompts_status_created` ON `prompts` (`status`,`createdAt`);--> statement-breakpoint
+CREATE INDEX `idx_prompts_callback` ON `prompts` (`status`,`callbackCompleted`,`callbackUrl`);--> statement-breakpoint
+CREATE INDEX `idx_prompts_status_priority_created` ON `prompts` (`status`,`priority`,`createdAt`);--> statement-breakpoint
 CREATE INDEX `idx_prompts_client_created` ON `prompts` (`clientName`,`createdAt`);--> statement-breakpoint
 CREATE UNIQUE INDEX `idx_prompts_client_request` ON `prompts` (`clientName`,`requestId`);